osd: EC optimizations: add written and present shard sets to pg_log_enty_t
Add two new sets to the pg_log_entry for use by EC optimization pools.
The written shards set tracks which shards were written to, the
present shards set tracks which shards were in the acting set at the
time of the write.

An empty set (default) is used to indicate all shards. For pools without
allow_ec_optimizations the written set is empty (indicating all shards are
written) and the present set is empty and unused.

Signed-off-by: Bill Scales <bill_scales@uk.ibm.com>