osd: add_next_event peering changes
add_next_event is used during peering to add a log entry to the local
shards copy of the log. With EC optimized pools not every log entry
has to be retained by every shard. During peering the primary sends
a full log to the replica which then needs to decide which log
entries it needs to add.

Signed-off-by: Bill Scales <156200352+bill-scales@users.noreply.github.com>