osd: TEMP primary shard selection
Temporary code to restrict the selection of the primary shard,
blocking any of the nonprimary shards from being selected.

This code works but because it changes OSDMap it requires all
OSDs and all clients to be running the new code. We want to do
better than this and only require all OSDs to be running the
new code, therefore need to rewrite this to use pgtemp instead.

Signed-off-by: Bill Scales <156200352+bill-scales@users.noreply.github.com>