mgr/smb: Add ceph_snapshots vfs module to share definition
In order to expose and manage snapshots as previous versions from
Windows we make use of the existing ceph_snapshots vfs module from
Samba. vfs_ceph_snapshots by default assumes the snap directory to
be named ".snap" with an option to configure a different name to
match the value for 'client_snapdir' parameter for CephFS client.
We follow the default behaviour to avoid further complications on
a live running cluster and thereby recommend to keep the default
for 'client_snapdir'. This limitation may slightly change in future
to dynamically detect the current snap directory name while shares
are created and not afterwards.

ref: https://www.samba.org/samba/docs/current/man-html/vfs_ceph_snapshots.8.html

Signed-off-by: Anoop C S <anoopcs@cryptolab.net>