mgr/smb: add password filter support to ceph smb apply command
The apply command both takes passwords in the input and can display them
in the results. Add two options --password-filter and
--password-filter-out that control how these passwords are filtered. If
--password-filter is given but --password-filter-out is not, the output
value is assumed to be the same as the input. So for example you could
run `ceph smb apply -i /file --password-filter=base64` which would
assume the password values in the input are base64 encoded and then
the results would include the same values.
`ceph smb apply -i /file --password-filter-out=hidden` would take
a file with unobscured passwords but emit results with passwords
hidden.

Signed-off-by: John Mulligan <jmulligan@redhat.com>