script/run-make: invoke clang-19 install via install-deps.sh
Replace the code to download and run llvm.sh with an argument to invoke
the similar functionality to install-deps. Note that in order not to
add as few special cases to install-deps as possible, the we are
enabling the llvm apt repo but the existing  `INSTALL_EXTRA_PACKAGES`
still is used to trigger the install of clang.

Signed-off-by: John Mulligan <jmulligan@redhat.com>