script/buildcontainer-setup: set JENKINS_HOME while building image
Set the JENKINS_HOME environment variable while building the builder
image. This is needed because parts of scripts like run-make.sh and
install-deps.sh key off of this variable. Since we want to be able to
use the build container to build, run "make check" and the like, we want
that environment to be as similar to the jenkins CI environment as we
can make it.

Signed-off-by: John Mulligan <jmulligan@redhat.com>