qa/workunits/rbd: fix positional argument expansion in create_image()
Sticking $@ into a string that is supposed to form a command isn't
right because the string would be broken apart when $@ has more than
one argument:

  If the double-quoted expansion occurs within a word, the expansion of
  the first parameter is joined with the beginning part of the original
  word, and the expansion of the last parameter is joined with the last
  part of the original word.

Resort to $* despite its shortcomings -- given run_cmd() signature it's
the only practical fixup.  A wrapper such as run_cmd() should really be
variadic or take an array instead of insisting on a single string.

This unbreaks "TEST: data pool".

Signed-off-by: Ilya Dryomov <idryomov@gmail.com>