crimson/osd/pg: Logically ignore older repops replies
When recieving a reply from peers we:
1) Update our local peer (from pending_trans) last_complete_ondisk
2) Update PeeringState::peer_last_complete_ondisk
3) If all peers replied:
     complete_write- upate PeertingState::pg_committed_to/last_complete_ondisk

In Crimson, we seperate between the handling of each write to the
peer replies awaiting. That means that the following scenario is
possible:

1) Handle write to obj a version 1'1
2) Append write 1'1 to pglog
3) Share write (1'1) with replicas
4) Write to obj a version 1'2
5) Append write 1'2 to pglog
6) Share write (1'2) with replicas
7) Peer replies from write (1'2) recieved.
8) Peer replies from write (1'1) recieved.

In this case, it's safe to ignore the peer replies
from write (1'1) when updating our peer lcod and this pg
pct and lcod (since we already have a later version commited).

With this fix, we check if the reply recieved can be ignored
before updating this pg peer and state.

Fixes: https://tracker.ceph.com/issues/69439

Signed-off-by: Matan Breizman <mbreizma@redhat.com>