pybind/mgr/status: Do not assert on empty metadata
When a OSD is marked as down and out, calling

```
$ ceph osd status
Error EINVAL: Traceback (most recent call last):
  File "/usr/share/ceph/mgr/mgr_module.py", line 1864, in _handle_command
    return CLICommand.COMMANDS[cmd['prefix']].call(self, cmd, inbuf)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/share/ceph/mgr/mgr_module.py", line 499, in call
    return self.func(mgr, **kwargs)
           ^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/share/ceph/mgr/status/module.py", line 337, in handle_osd_status
    assert metadata
AssertionError
```

will crash with the message above. The error message in
`ActivePyModules::get_metadata_python` also suggest that this method can
return `None` if the service is missing.

The assert was introduced in 5ac2901f54ff4c0b1702e1b82c5139b3d5baf39a.
In Reef the metadata would contain the hostname where the OSD is hosted
but in Squid there is no metadata.

With this patch we get the following output:

```
$ ceph osd status
41  myhost  1421M  30.6G      0        0       0        0   exists,up
42          0      0          0        0       0        0   exists
```

Signed-off-by: Maximiliano Sandoval <m.sandoval@proxmox.com>