rgw: modify radoslist to better support the rgw-gap-list tool
When the `radosgw-admin bucket radoslist ...` sub-command was
introduced, it was written specifically for finding orphans. It has
since been updated to work for finding gaps, that is indexed RGW
objects that are missing one or more supporting rados objects.

When a head object was not found, it was ignored. Now it does produce
output with the oid and related information for the missing head
object.

Signed-off-by: J. Eric Ivancich <ivancich@redhat.com>