mgr: rename APIs for fetching perf schema and perf counter values to indicate that they only work on unlabeled counters.
The current API's for fetching performance schema only fetches unlabeled
counters and the API for fetching the latest value of the performance
counters also only works on unlabeled counters. We update the name of
these API's to correctly reflect this behaviour.

This change is also required to make sure that we do not break any
existing functionality while we work on the support of gradually phasing
all the manager clients such as telemetry to work with labeled counters.

Signed-off-by: Naveen Naidu <naveen.naidu@ibm.com>