mgr: add get_perf_counter API to fetch the values of labeled counters
get_perf_counter API loops through the labeled counters present in the
schema that is fetched via get_perf_schema and fetches the latest value
of these counters.

The logic is similar to get_unlabeled_perf_counter, with the only
difference being is that we use the labeled API variants to fetch the
latest value.

Signed-off-by: Naveen Naidu <naveen.naidu@ibm.com>