test/common: skip google tests which create core dumps in test_interval_set
CI Pipelines are being broken because this test is creating a number of core dumps. This
does not make the test fail, but it does create some core dumps. This appears to be
breaking something in the CI pipeline.  This commit is a workaround and I will find a
better solution later.

Fixes: https://tracker.ceph.com/issues/70543

Signed-off-by: Alex Ainscow <aainscow@uk.ibm.com>