osd: Restrict choice of primary shard for ec_optimizations pools
Pools with ec_optimizations enabled have restrictions on which
shards are permitted to become the primary because not all shards
are updated for every I/O.

To preserve backwards compatibility with downlevel clients
pg_temp is used to shuffle the order of the shards so that
the exist OSDMap code selects an appropraite primary. The
shuffling of pg_temp occurs in OSDMonitor when pg_temp is
set, this is reverted in PeeringState when setting the acting
set after OSDMap has selected the primary.

FIXME: This commit undoes changes made by an earlier attempt
to change the primary selection. These commits need to be
squashed together.

Signed-off-by: Bill Scales <bill_scales@uk.ibm.com>