qa/cephfs: raise exception when expected progress events are not found
Method filter_in_only_clone_pevs() looks for progress events with ID
"mgr-vol-ongoing-clones" or "mgr-vol-total-clones". And when it doesn't
find either of them, it returns empty dictionary instead of raising
exception or returning a value indicating this error.

This is problematic because if somehow one of these progress events IDs
is accidentally and undesirably renamed in the source code, tests won't
catch this issue. This causes the tests to fail at a much later point
(which makes it harder to debug) or probably cause tests to run
successfully without exposing this bug, which is even worse.

Therefore, when neither of these progress IDs are found raise an
exception and cause the test to abort.

Signed-off-by: Rishabh Dave <ridave@redhat.com>