qa/cephfs: improve filter_in_only_clone_pevs
filter_in_only_clone_pevs should return None if variable progress_events
received by it are empty.

And raise an exception if the progress_events is not a dictionary since
if it is not a dictionary (of progress events from "ceph status" command
output) it makes no sense to proceed.

Signed-off-by: Rishabh Dave <ridave@redhat.com>