src/common: add guidance for deep-scrubbing ratio warning
Add an explanation of how to set the value of
"mon_warn_pg_not_deep_scrubbed_ratio" to the confval definition of that
variable. Although this variable contains the string "mon", it is set on
the Manager. I have added a note to direct users to set this value on
the Manager.

This issue was pointed out by Petr Tlapa on Slack in late March of 2025.

Signed-off-by: Zac Dover <zac.dover@proton.me>