mgr: add status command
The Rook operator relies on a `status` (asok) command to the mgr to verify it
is operational. However, this command was actually served by the CephFS
`Client` library that the ceph-mgr daemon statically linked in and used for
sending MDS commands. That statically linked client was removed via 8aaa9e46
(case insensitive directory tree feature).

So, add a legitimate ceph-mgr status command which simply outputs an empty
dictionary and returns success (0). TBD on adding useful information.

Fixes: 8aaa9e4613b7f310a365f86fbdee65c828784f75
Fixes: https://tracker.ceph.com/issues/70571
Signed-off-by: Patrick Donnelly <pdonnell@ibm.com>