crimson/os/seastore/cache: adjust transactions' read_set when completing transaction commits
With LBACursor, we can construct lba iterators in a bottom-up way, which
introduce a new case in which a transaction may hit an initial pending
lba node that another transaction is committing.

In this case, the hitting transaction add that initial pending extent
into its read_set to be able to respond to the extent invalidation
events. This make it necessary for the hitted transaction to adjust the
hitting transaction's read_set for the initial pending extent in
question, as its paddr is changed from a relative one to an absolute
one.

Signed-off-by: Xuehan Xu <xuxuehan@qianxin.com>