libcephfs_proxy: workaround for delays and hangs in the proxy
When ceph_ll_lookup() is done on ".", it returns a new reference to the
parent inode itself. However this may cause some delays or problems on
disconnect (see https://tracker.ceph.com/issues/70553).

However, if "." is replaced by "", we still get a new reference to the
parent inode (which is what we really need), without causing any side
effect.

This patch uses this alternative in the libcephfs proxy to avoid the
problem.

Signed-off-by: Xavi Hernandez <xhernandez@gmail.com>