test/osd: Adapt strategy for testing asserts in interval_sets
Previously the strategy for testing interval_sets was to use a gtest
feature called "EXPECT_DEATH" This is, however, slow and creates
core dumps. Instead, we use a mechanism which replaces the
ceph_assert macro with a strict_assert, which behaves the same in
product code, but in the unit test can be tested using the faster
EXPECT_THROW() mechanism in gtest.

Signed-off-by: Alex Ainscow <aainscow@uk.ibm.com>