cephadm/cephadmlib/daemons/ceph.py: increase aio-max-nr
Previous work increased the aio-max-nr to 1048576 [1].
Crimson-osd might need this value to be larger in order
to deploy correctly.
In future work, if this turns out to be tricky, we can update
cephadm to make use of `90-ceph-osd.conf` file which changes between
build types. See possible usage in sysctl.py.
For now, increasing the value for any Class OSD should work.

[1] https://github.com/ceph/ceph/pull/41264

Signed-off-by: Matan Breizman <mbreizma@redhat.com>