cephadm: stop sidecar systemd units when restarting main units
Previously, sidecar units might (will) not be stopped when the primary
unit is stopped. This breaks the normal cephadm workflow to redeploy:
stop and then start the services. This patch reuses the logic from
the service stop to stop the individual sidecars.

Note: the sidecars are not individually started as that should already
be handled automatically when starting the primary service!

Co-authored-by: Adam King <adking@redhat.com>
Signed-off-by: Adam King <adking@redhat.com>
Signed-off-by: John Mulligan <jmulligan@redhat.com>