crimson/osd/../client_request: can_serve_replica_read after getting obc
When handling a client request on a clone object, m->get_hobj() might
not actually represent the "actual" object from the OSD's prespective.
See: ObjectContext::resolve_oid for how this logic is handled.

For this reason, we should move the can_serve_replica_read call on the
object only *after* it was resolved (in load_and_lock).

Fixes (1/3): https://tracker.ceph.com/issues/70639

Signed-off-by: Matan Breizman <mbreizma@redhat.com>