src/common: add guidance for mon_warn_pg_not_scrubbed
Add an explanation of how to set the value of
"mon_warn_pg_not_scrubbed_ratio" to the confval definition of that
variable. Although this variable contains the string "mon", it is set on
the Manager. I have added a note to direct users to set this value on
the Manager.

This issue was pointed out by Petr Tlapa on Slack in late March of 2025.

This issue is part of a small project that also encompasses
https://github.com/ceph/ceph/pull/62459, which is associated with commit
b6a1e04b746fc4309fce6e83c26f24fa82a9613f

Co-authored-by: Anthony D'Atri <anthony.datri@gmail.com>
Signed-off-by: Zac Dover <zac.dover@proton.me>