cmake/modules/BuildBoost: Fallback to gcc
See the following error when using clang-19:
```
A C++11 capable compiler is required for building the B2 engine.
Toolset 'clang' does not appear to support C++11.
```

Fallback to gcc toolset intead and don't set cxx as fallbacks:
```
For any toolset you can override the path to the compiler with the '--cxx'
option. You can also use additional flags for the compiler with the
'--cxxflags' option.

A special toolset; cxx, is available which is used as a fallback when a more
specific toolset is not found and the cxx command is detected. The 'cxx'
toolset will use the '--cxx' and '--cxxflags' options, if present.
```

Signed-off-by: Matan Breizman <mbreizma@redhat.com>