qa/cephfs: update TestPausePurging
This test has helper methods that construct paths of subvolumes after
they've been moved to trash directory. For this, subvolume path is
obtained before deletion by running "ceph fs subvolume getpath" command
and then manipulating the string obtained in the output of this command
to get the path of the subvolume after it has been moved to trash
directory.

These 2 specific string modifications are causing tests failures:
subvolume name is deleted from the "path" string and UUID is found and
added to this "path" string. This specific code is now outdated due to
introduction of subvolume v3. Revise and update it so that tests don't
fail due to introduction of subvolume v3.

Signed-off-by: Rishabh Dave <ridave@redhat.com>