qa/cephfs: update test_pausing_halts_ongoing_cloning
This test constructs path for clone subvolume using path of the
subvolume the snapshot of which serves as the source of cloning
operation. This code is now outdated due to introduction of subvolume
v3 since subvolume v3 has 4 more level of directories
(<subvol-name>/roots/<uuid>/mnt) instead of 2 (<subvol-name>/uuid>).
Update the test code according to this so that it doesn't fail.

Signed-off-by: Rishabh Dave <ridave@redhat.com>