common: refactor perf counter class
This refactor is done so as to be able to fetch the perf counters
via client APIs (say, libcephfs). The current mechanism of exporting
perf counters is via formatters (JSON, etc..) makes it a little
weird to expose the perf counters via a library API. This change
adds a mechanism to get perf counters by type packed in a struct.

Signed-off-by: Venky Shankar <vshankar@redhat.com>