mgr/vol: handle case where path goes missing for a clone
A thread is spawned to get the value of a certain extended attribute to
generate the progress statistics for the ongoing clone operations. In
case source and/or destination path for a clone operation goes missing,
this thread crashes. Instead of crashing, handle this case gracefully.

Fixes: https://tracker.ceph.com/issues/71019
Signed-off-by: Rishabh Dave <ridave@redhat.com>
(cherry picked from commit 359277bc5323823a5f86f368f4a3c56ed4fca404)