rgw/multisite: allow replication of non-versioned objects to versioned buckets
When a bucket transitions from non-versioned to versioned, there
may be a delay in synchronizing this state and the already ongoing
data replication of non-versioned objects. During this period, some
older non-versioned objects may still need to be replicated to the
now-versioned destination bucket.

To handle this, non-versioned objects can be replicated to a
versioned bucket with a null version, ensuring no data is lost
during the transition. However, replication should still skip
versioned objects when syncing to a non-versioned bucket.

Signed-off-by: Seena Fallah <seenafallah@gmail.com>