rbd_mirror: fix cross namespace group snapshot mirroring
In case if group belongs to non-default namespace, the mirror daemon on
secondary is setting an empty mirror peer uuid on non primary demote snapshot.
And when previous secondary turns into primary (by an explicit promote request)
the previous i.e. non primary demote snapshot is getting unlinked as part of
the promote request as there is no mirror peer uuid set on it. Because the
previous dependent snapshot is removed, there are split-brain errors leading
to recent test failures.

This fix makesure to set the right mirror peer uuid on the non primary demote
snapshot even if the group belong to non-default namespace.

Signed-off-by: Prasanna Kumar Kalever <prasanna.kalever@redhat.com>