mgr/vol: include group name in subvolume's pool namespace name
Include subvolume group name along with subvolume name in the pool
namepsace string for the subvolume. IOW, include it in the string that
serves as the value of the extended attribute
"ceph.dir.layout.pool_namespace", which is set for a subvolume during
its creation.

Right now, due to not including it, two different subvolumes with same
name but located in different subvolume groups undesirably end up in the
same namespace.

Fixes: https://tracker.ceph.com/issues/70668
Signed-off-by: Rishabh Dave <ridave@redhat.com>
(cherry picked from commit 45f8abab8f467043823241233664789f127a7c6b)