qa/cephfs: test that group name is included in pool namespace name
Test that subvolume group name is included in the string that serves as
the name of pool namespace for a subvolume. IOW, that serves as value of
the extended attribute "ceph.dir.layout.pool_namespace" for a subvolume.

Signed-off-by: Rishabh Dave <ridave@redhat.com>
(cherry picked from commit db8c3f99f97c3a8a332275ddee97d8303406e663)