common: md_config_cacher_t: add a method to force-update the cached values
required when the owner of the cacher object is accessing the
configuration value, in a call sequence originated by some
other handler (in the specific case that
triggered this here change, it was from the OSD::handle_conf_change()).
This allows the owner to force an update of the cached
values before using them.

Signed-off-by: Ronen Friedman <rfriedma@redhat.com>