bluestore: disable Elastic Shared Blobs temporarily
It's been reported by several users that the ESB doesn't
to work well and OSDs crash when enabled.

The bug is tracked in https://tracker.ceph.com/issues/70390

This is a temporary change in the default to avoid users
being caught up inadvertently as recreating all the OSDs
could be very painful.

Related to: https://tracker.ceph.com/issues/70390

Signed-off-by: Ponnuvel Palaniyappan <pponnuvel@gmail.com>