cephadm: check "ceph_device_lvm" field instead of "ceph_device" during zap
This field was renamed as part of
https://github.com/ceph/ceph/commit/4941d098e337f2b7ad8c6f7c90be3ae252d22f7b
but the cephadm piece was still looking for "ceph_device" meaning
we would never zap any devices. Worth noting this won't
get caught by teuthology as we make OSDs in that env directly
on top of LVs made out of a split up nvme drive and
ceph-volume inventory doesn't list LVs.

Signed-off-by: Adam King <adking@redhat.com>