mgr/cephadm: fix staggered upgrade daemon type warning
This was meant to just be daemons that use the ceph image,
which everything under "gateway" daemons used to be before
the addition of nvmeof and smb. This patch updates it to
explicitly reference the daemon types that use the ceph image.

Signed-off-by: Adam King <adking@redhat.com>
(cherry picked from commit 2d248199cda0257884d2aac6a5003118e86c5880)