mgr/vol: make "subvolume ls" command compatible with v3
Until v2, all directories under group directory were listed and reported
as it is when "subvolume ls" command was run. But in v3, this alone is
not sufficient since the subvol directory may exist but it may contain
only previous incarnations. IOW, its current incarnation does not exist
right now. And in that case it makes no sense to report it.

Signed-off-by: Rishabh Dave <ridave@redhat.com>