doc/releases: Fix invalid triple backticks in reef.rst squid.rst
Triple backtick does not create a code block in RST,
instead it renders as an inline code with the third
backtick rendered as-is.
This makes newlines in multiline code merged to a
single line and it makes the whole thing nonsense.

Change the second intended code block to use a code
block with a bash prompt.

Signed-off-by: Ville Ojamo <14869000+bluikko@users.noreply.github.com>