rgw/admin: Add max-entries and marker to bucket list
This adds pagination to the /admin/bucket endpoint for the
Admin API.

If a user has a lot of buckets the /admin/bucket endpoint
that is listing buckets can be so long that the HTTP request
gets a timeout.

This adds the ``max-entries`` and ``marker`` query parameters
to the API to support pagination. If ``max-entries`` is given
we introduce a new format for the HTTP response body the same
way that metadata API does, if it's not given we return the
response with the same body as before and thus retaining the
backward compatibility of the API.

This adds a Python3 based test suite that tests all of this
functionality to verify the behaviour and the HTTP response
body itself.

This fixes the pagination mentioned in tracker [1] and thus
fixes (or atleast partially fixes) that.

[1] https://tracker.ceph.com/issues/22168

Fixes: https://tracker.ceph.com/issues/22168
Signed-off-by: Tobias Urdin <tobias.urdin@binero.com>
(cherry picked from commit 1d5523ec0bec916e0a87fdcb8d27b67753e477b6)