libcephfs_proxy: fully initialize async I/O structure
The ceph_ll_io_info structure has recently been extended to support
zerocopy operations. The proxy was initializing just the known members,
so, after the zerocopy support, it was passing garbage in some fields,
causing failures.

This patch completely clears the whole structure to be sure that
everything is initialized to its default value.

Signed-off-by: Xavi Hernandez <xhernandez@gmail.com>