osd/scrub: a single counters selection mechanism - step 1
Following the preceeding PR, the Scrubber now employs
two methods for selecting the specific subset of performance
counters to update (the replicated pool set or the EC one).
The first method is using labeled counters, with 4 optional labels
(Primary/Replica X Replicated/EC Pool). The second method
is by naming the specific OSD counters to use in ScrubIoCounterSet
objects, then selecting the appropriate set based on the pool type.

This commit is the first step on the path to unifying the two
methods - discarding the use of labeled counters, and only
naming OSD counters.

Signed-off-by: Ronen Friedman <rfriedma@redhat.com>