mgr/vol: handle exception when path for lstat() goes missing
It might happen that clone index entry goes missing because a clone job
was completed or cancelled. In such a case, lstat() to clone entry's
path would fail. Catch the exception in such a case and handle it so
that clone progress reporter thread doesn't crash.

Crashing of clone progress reporter thread causes clone progress bars to
not be removed from "ceph status" output when they should, resulting in
these bars to being in stuck in the output forever.

Fixes: https://tracker.ceph.com/issues/70941
Signed-off-by: Rishabh Dave <ridave@redhat.com>