mgr/vol: don't delete user-created pool in "volume create" command
If one of the pool names passed to "ceph fs volume create" command
(through --data-pool and --meta-pool name) is absent, don't delete the
pool that is present and passed to this command during the cleanup code
of this command.

IOW, "volume create" command should continue deleting pool created by it
but not delete pool created by the user.

Fixes: https://tracker.ceph.com/issues/70945
Signed-off-by: Rishabh Dave <ridave@redhat.com>