osd: Fix Truncates in Optimised EC
The previous truncate code attempted to perform a non-aligned truncate by
creating a zero buffer at the end of the object, which was written.

The new code initially truncates to the exact size of the user object before
growing the object to the required 4k alignment. This simpler arrangement
also simplifies the rollback.

Signed-off-by: Alex Ainscow <aainscow@uk.ibm.com>