test/libcephfs: copy DT_NEEDED entries from input libraries
On Ubuntu 22.04, the linker is not stumbling thinking the libceph-common
library is missing on the command-line. This appears to be a bug and the only
workaround I've found is to copy the DT_NEEDED entries for the input shared
objects (which is traditional linker behavior). I don't have an explanation for
why this occurs only for a few test executables.

Fixes: https://tracker.ceph.com/issues/70498
Signed-off-by: Patrick Donnelly <pdonnell@ibm.com>
(cherry picked from commit c6ac71d398c16afa4aa888feaf7087bf61038bf0)