crimson/osd/osd_operations/pg_splitting: make use of PGCreationBlockingEvent
For every split child PG being created, make an entry in the wait_for_pg map using
a PGCreationBlockingEvent trigger.
This change ensures that even split PGs are part of the pgs_creating map that tracks
PGCreationState of a PG.
We no longer need a separate workflow for split PGs in the pg_created function.

Signed-off-by: Aishwarya Mathuria <amathuri@redhat.com>